+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 v0.2.0 (2023-12-28)
 --------------------
 *   Exceptions for specific errors.
+*   ``IpernityAPI.call()`` raises ``APIRequestError`` instead of ``HTTPError``
+    if HTTP request fails.
 *   Removed obsolete ``_replace_file`` method.
 
 v0.1.5 (2023-12-09)